// src/types/bank.types.ts
export interface Bank {
    id: number;
    name: string;
    code: string;
    bin: string;
    shortName: string;
    logo: string;
    transferSupported: number;
    lookupSupported: number;
    short_name: string;
    support: number;
    isTransfer: number;
    swift_code: string;
}

export interface VietQRResponse {
    code: string;
    desc: string;
    data: Bank[];
}